Update README.md
A project on Lucky Draw Winner Selection. It is prepared by 1. Amrina Rahman (242-35-589) 2.Al Jubair Badhon (242-35-428) 3. Milton (242-35--589) .
This project is a simple lucky draw selector that randomly picks a winner from a list of contestants. Users can input multiple names, and the program will use a random selection method to determine the lucky winner. This is useful for giveaways, raffles, or any event where a fair and unbiased selection is needed.
 The code is designed to be user-friendly, providing clear prompts and feedback throughout the process.